Move lakeroad import stuff into its own function
