+++ lakeroad-egglog/src/lib.rs
@@ -1,6 +1,14 @@
-use std::collections::HashSet;
+use std::{
+    collections::{HashMap, HashSet},
+    sync::Arc,
+};
 
-use egglog::{ast::Literal, Term, TermDag};
+use egglog::{
+    ast::{Literal, Symbol},
+    constraint::{SimpleTypeConstraint, TypeConstraint},
+    sort::{FromSort, I64Sort, IntoSort, Sort, VecSort},
+    EGraph, PrimitiveLike, Term, TermDag, Value,
+};
 
 pub fn to_verilog(term_dag: &TermDag, id: usize) -> String {
     // let mut wires = HashMap::default();
@@ -203,6 +211,106 @@ pub fn to_verilog(term_dag: &TermDag, id: usize) -> String {
         registers = registers,
         module_declarations = module_declarations,
     )
+}
+
+/// Import Lakeroad language into an EGraph.
+///
+/// TODO(@gussmith23): Ideally, this would be done via an `import` statement.
+/// That's not currently possible because of the Rust-defined primitive
+/// `debruijnify` in Lakeroad.
+pub fn import_lakeroad(egraph: &mut EGraph) {
+    egraph
+        .parse_and_run_program(
+            r#"
+(include "egglog_src/lakeroad-antiunify.egg")
+    "#,
+        )
+        .unwrap();
+
+    struct DeBruijnify {
+        in_sort: Arc<VecSort>,
+        out_sort: Arc<VecSort>,
+        i64_sort: Arc<I64Sort>,
+    }
+
+    impl PrimitiveLike for DeBruijnify {
+        fn name(&self) -> Symbol {
+            "debruijnify".into()
+        }
+
+        fn get_type_constraints(&self) -> Box<dyn TypeConstraint> {
+            Box::new(SimpleTypeConstraint::new(
+                self.name(),
+                vec![self.in_sort.clone(), self.out_sort.clone()],
+            ))
+        }
+
+        fn apply(&self, values: &[crate::Value], egraph: &EGraph) -> Option<crate::Value> {
+            let in_vec = Vec::<Value>::load(&self.in_sort, &values[0]);
+
+            let mut seen_values: HashMap<Value, i64> = HashMap::new();
+            let mut next_id = 0;
+            let mut out = vec![];
+
+            for value in in_vec {
+                // Get representative value.
+                let value = egraph.find(value);
+
+                // If we haven't assinged it a number yet, give it the next one.
+                if !seen_values.contains_key(&value) {
+                    seen_values.insert(value, next_id);
+                    next_id += 1;
+                }
+
+                // Add the number to the output vector.
+                out.push(seen_values[&value].store(&self.i64_sort).unwrap());
+            }
+
+            out.store(&self.out_sort)
+        }
+    }
+
+    egraph.add_primitive(DeBruijnify {
+        i64_sort: egraph.get_sort().unwrap(),
+        in_sort: egraph
+            .get_sort_by(|s: &Arc<VecSort>| s.name() == "ExprVec".into())
+            .unwrap(),
+        out_sort: egraph
+            .get_sort_by(|s: &Arc<VecSort>| s.name() == "IVec".into())
+            .unwrap(),
+    });
+
+    egraph
+        .parse_and_run_program(
+            r#"
+;;; Op2
+;; hole, hole
+(rewrite
+  (Op2 op e1 e2)
+  (apply (MakeModule (Op2_ op (Hole) (Hole)) (debruijnify (vec-of e1 e2))) (vec-of e1 e2))
+  :ruleset enumerate-modules
+)
+;; merge, hole
+(rewrite
+  (Op2 op (apply (MakeModule graph indices) args) e)
+  (apply (MakeModule (Op2_ op graph (Hole)) (debruijnify (vec-push args e))) (vec-push args e))
+  :ruleset enumerate-modules
+)
+;; hole, merge
+(rewrite
+  (Op2 op e (apply (MakeModule graph indices) args))
+  (apply (MakeModule (Op2_ op (Hole) graph) (debruijnify (vec-append (vec-of e) args))) (vec-append (vec-of e) args))
+  :ruleset enumerate-modules
+)
+;; merge, merge
+(rewrite
+  (Op2 op (apply (MakeModule graph0 indices0) args0) (apply (MakeModule graph1 indices1) args1))
+  (apply (MakeModule (Op2_ op graph0 graph1) (debruijnify (vec-append args0 args1))) (vec-append args0 args1))
+  :ruleset enumerate-modules
+)
+    "#,
+        )
+        .unwrap();
 }
 
 #[cfg(test)]
